CHANGE GIT HISTORY TO REMOVE VENDOR FILES FROM THE REPO
Before this commit, there was a src/ folder with java binary files (.jars) from the HTMLUnit package. Those were really big files and were making the repo grow a lot. I removed the src folder from git's history and initialized a submodule pointing to (https://github.com/Nerian/html-unit-vendor). As a result this repo won't be compatible with current Akephalos's forks, but since this is the most updated one, that won't be a problem. Just fork again.